Unify multi-tenant behavior tests to save time during setup/create schema
