+++ tcms_tenants/tests/test_api.py
@@ -11,13 +11,10 @@ from mock import patch
 
 from django.contrib.auth import get_user_model
 from django.test import override_settings
-from django.utils import timezone
 
 from django_tenants import utils
-from parameterized import parameterized
-from tcms.tests.factories import ProductFactory
 
-from tcms_tenants.tests import TenantGroupsTestCase, UserFactory
+from tcms_tenants.tests import TenantGroupsTestCase
 from tenant_groups.models import Group as TenantGroup
 
 UserModel = get_user_model()
@@ -129,112 +126,4 @@ class TenantInviteApiTestCase(TenantGroupsTestCase):
         self.assertEqual(
             90,
             UserModel.objects.filter(username__startswith="invited-silently").count(),
-        )
-
-
-class ComponentCreateApiTestCase(TenantGroupsTestCase):
-    @classmethod
-    def setUpClass(cls):
-        super().setUpClass()
-
-        # Create users user_01..user_05 in public schema
-        cls.user_01 = UserFactory()
-        cls.user_02 = UserFactory()
-        # user_03 is not authorized for either tenant
-        cls.user_03 = UserFactory()
-        cls.user_04 = UserFactory()
-        cls.user_05 = UserFactory()
-
-        # Setup self.tester and users 01, 05 for self.tenant
-        with utils.tenant_context(cls.tenant):
-            TenantGroup.objects.get(name="Tester").user_set.add(cls.tester)
-            cls.tenant.authorized_users.add(cls.user_01)
-            cls.tenant.authorized_users.add(cls.user_05)
-
-        # Create second tenant
-        with utils.schema_context("public"):
-            ts = int(timezone.now().timestamp())
-            cls.tenant2 = utils.get_tenant_model()(
-                schema_name=f"tenant_{ts}",
-                owner=cls.user_02,
-            )
-            cls.tenant2.save()
-
-            cls.domain2 = utils.get_tenant_domain_model()(
-                tenant=cls.tenant2, domain=f"second-{ts}.example.com"
-            )
-            cls.domain2.save()
-
-        # User 02 is the owner of tenant2, user 04 is also authorized
-        with utils.tenant_context(cls.tenant2):
-            cls.tenant2.authorized_users.add(cls.user_02)
-            cls.tenant2.authorized_users.add(cls.user_04)
-
-        cls.product = ProductFactory()
-
-    @parameterized.expand(
-        [
-            ("user_01", lambda self: self.user_01),
-            ("user_05", lambda self: self.user_05),
-            ("tester", lambda self: self.tester),
-            ("tenant.owner", lambda self: self.tenant.owner),
-        ]
-    )
-    def test_component_create_with_authorized_user(self, _, get_user):
-        initial_owner = get_user(self)
-
-        response = self.client.post(
-            "/json-rpc/",
-            {
-                "id": "jsonrpc",
-                "jsonrpc": "2.0",
-                "method": "Component.create",
-                "params": [
-                    {
-                        "name": f"Component for {initial_owner.username}",
-                        "product": self.product.pk,
-                        "initial_owner": initial_owner.pk,
-                    }
-                ],
-            },
-            content_type="application/json",
-        )
-        self.assertEqual(HTTPStatus.OK, response.status_code)
-        data = json.loads(response.content)
-        self.assertIn("result", data)
-        component = data["result"]
-        self.assertEqual(component["name"], f"Component for {initial_owner.username}")
-        self.assertEqual(component["initial_owner"], initial_owner.pk)
-
-    @parameterized.expand(
-        [
-            ("user_02", lambda self: self.user_02),
-            ("user_03", lambda self: self.user_03),
-            ("user_04", lambda self: self.user_04),
-        ]
-    )
-    def test_component_create_with_unauthorized_user(self, _, get_user):
-        initial_owner = get_user(self)
-
-        response = self.client.post(
-            "/json-rpc/",
-            {
-                "id": "jsonrpc",
-                "jsonrpc": "2.0",
-                "method": "Component.create",
-                "params": [
-                    {
-                        "name": f"Component for {initial_owner.username}",
-                        "product": self.product.pk,
-                        "initial_owner": initial_owner.pk,
-                    }
-                ],
-            },
-            content_type="application/json",
-        )
-        self.assertEqual(HTTPStatus.OK, response.status_code)
-        data = json.loads(response.content)
-        self.assertIn("error", data)
-        self.assertRegex(
-            data["error"]["message"], r"initial_owner.*Select a valid choice"
         )
